Bug 1715134 - Make form control type an enum class. r=masayuki
Differential Revision: https://phabricator.services.mozilla.com/D117087

UltraBlame original commit: 201320cde51a0c5222cdb7654b7f56b227b9d4fd

+++ dom/base/nsContentUtils.cpp
@@ -3044,7 +3044,7 @@ void nsContentUtils::GenerateStateKey(nsIContent* aContent, Document* aDocument,
       }
 
       
-      KeyAppendInt(control->ControlType(), aKey);
+      KeyAppendInt(int32_t(control->ControlType()), aKey);
 
       
       HTMLFormElement* formElement = control->GetFormElement();
